Remove obsolete overload of AuthenticationExtensions.AddJwtBearerAuthenticationForWebApp
