+++ source/App/source/ExampleHost.WebApi04/Startup.cs
@@ -12,11 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Core.App.Common.Extensions.Options;
 using Energinet.DataHub.Core.App.WebApp.Extensions.Builder;
 using Energinet.DataHub.Core.App.WebApp.Extensions.DependencyInjection;
 using ExampleHost.WebApi04.Security;
-using Microsoft.IdentityModel.Protocols.Configuration;
 
 namespace ExampleHost.WebApi04;
 
@@ -33,11 +31,12 @@ public class Startup
     {
         services.AddControllers();
 
+        // Http => Authentication (verified in tests)
         // Configure for testing
         AuthenticationExtensions.DisableHttpsConfiguration = true;
-
-        AddJwtAuthentication(services);
-        services.AddUserAuthenticationForWebApp<ExampleSubsystemUser, ExampleSubsystemUserProvider>();
+        services
+            .AddJwtBearerAuthenticationForWebApp(Configuration)
+            .AddUserAuthenticationForWebApp<ExampleSubsystemUser, ExampleSubsystemUserProvider>();
     }
 
     public void Configure(IApplicationBuilder app, IWebHostEnvironment environment)
@@ -54,25 +53,5 @@ public class Startup
         {
             endpoints.MapControllers();
         });
-    }
-
-    /// <summary>
-    /// Here we configure the application using the "old" (obsolete) extension, for testing scenarious where we don't have nested tokens.
-    /// </summary>
-    protected virtual void AddJwtAuthentication(IServiceCollection services)
-    {
-        var authenticationOptions = Configuration
-            .GetRequiredSection(UserAuthenticationOptions.SectionName)
-            .Get<UserAuthenticationOptions>();
-
-        if (authenticationOptions == null)
-            throw new InvalidConfigurationException("Missing authentication configuration.");
-
-#pragma warning disable CS0618 // Type or member is obsolete
-        services.AddJwtBearerAuthenticationForWebApp(
-            authenticationOptions.ExternalMetadataAddress,
-            authenticationOptions.InternalMetadataAddress,
-            authenticationOptions.BackendBffAppId);
-#pragma warning restore CS0618 // Type or member is obsolete
     }
 }
